Refactored to split RS endpoint logic to its own ReportStreamEndpointClient class
