+++ etor/src/main/java/gov/hhs/cdc/trustedintermediary/external/reportstream/ReportStreamEndpointClient.java
@@ -1,0 +1,179 @@
+package gov.hhs.cdc.trustedintermediary.external.reportstream;
+
+import gov.hhs.cdc.trustedintermediary.context.ApplicationContext;
+import gov.hhs.cdc.trustedintermediary.etor.orders.UnableToSendOrderException;
+import gov.hhs.cdc.trustedintermediary.wrappers.AuthEngine;
+import gov.hhs.cdc.trustedintermediary.wrappers.Cache;
+import gov.hhs.cdc.trustedintermediary.wrappers.HapiFhir;
+import gov.hhs.cdc.trustedintermediary.wrappers.HttpClient;
+import gov.hhs.cdc.trustedintermediary.wrappers.HttpClientException;
+import gov.hhs.cdc.trustedintermediary.wrappers.Logger;
+import gov.hhs.cdc.trustedintermediary.wrappers.MetricMetadata;
+import gov.hhs.cdc.trustedintermediary.wrappers.SecretRetrievalException;
+import gov.hhs.cdc.trustedintermediary.wrappers.Secrets;
+import gov.hhs.cdc.trustedintermediary.wrappers.formatter.Formatter;
+import gov.hhs.cdc.trustedintermediary.wrappers.formatter.FormatterProcessingException;
+import gov.hhs.cdc.trustedintermediary.wrappers.formatter.TypeReference;
+import java.time.LocalDateTime;
+import java.time.temporal.ChronoUnit;
+import java.util.Map;
+import java.util.Optional;
+import javax.annotation.Nonnull;
+import javax.inject.Inject;
+
+public class ReportStreamEndpointClient {
+    private static final String RS_URL_PREFIX_PROPERTY = "REPORT_STREAM_URL_PREFIX";
+    private static final String RS_DOMAIN_NAME =
+            Optional.ofNullable(ApplicationContext.getProperty(RS_URL_PREFIX_PROPERTY))
+                    .map(urlPrefix -> urlPrefix.replace("https://", "").replace("http://", ""))
+                    .orElse("");
+    private static final String RS_WATERS_API_URL =
+            ApplicationContext.getProperty(RS_URL_PREFIX_PROPERTY) + "/api/waters";
+    private static final String RS_AUTH_API_URL =
+            ApplicationContext.getProperty(RS_URL_PREFIX_PROPERTY) + "/api/token";
+    private static final String RS_HISTORY_API_URL =
+            ApplicationContext.getProperty(RS_URL_PREFIX_PROPERTY)
+                    + "/api/waters/report/{id}/history";
+
+    private static final String OUR_PRIVATE_KEY_ID =
+            "trusted-intermediary-private-key-" + ApplicationContext.getEnvironment();
+    private static final String RS_TOKEN_CACHE_ID = "report-stream-token";
+
+    private static final String CLIENT_NAME = "flexion.etor-service-sender";
+    private static final Map<String, String> RS_AUTH_API_HEADERS =
+            Map.of("Content-Type", "application/x-www-form-urlencoded");
+
+    @Inject private HttpClient client;
+    @Inject private AuthEngine jwt;
+    @Inject private Formatter formatter;
+    @Inject private HapiFhir fhir;
+    @Inject private Logger logger;
+    @Inject private Secrets secrets;
+    @Inject private Cache cache;
+
+    @Inject MetricMetadata metadata;
+
+    private static final ReportStreamEndpointClient INSTANCE = new ReportStreamEndpointClient();
+
+    public static ReportStreamEndpointClient getInstance() {
+        return INSTANCE;
+    }
+
+    private ReportStreamEndpointClient() {}
+
+    protected String getRsToken() throws UnableToSendOrderException {
+        logger.logInfo("Looking up ReportStream token");
+
+        var token = cache.get(RS_TOKEN_CACHE_ID);
+
+        if (token != null && isValidToken(token)) {
+            logger.logDebug("valid cache token");
+            return token;
+        }
+
+        token = requestToken();
+
+        cache.put(RS_TOKEN_CACHE_ID, token);
+
+        return token;
+    }
+
+    protected boolean isValidToken(String token) {
+        LocalDateTime expirationDate = jwt.getExpirationDate(token);
+
+        return LocalDateTime.now().isBefore(expirationDate.minus(15, ChronoUnit.SECONDS));
+    }
+
+    protected String sendRequestBody(@Nonnull String json, @Nonnull String bearerToken)
+            throws UnableToSendOrderException {
+        logger.logInfo("Sending payload to ReportStream");
+
+        String res = "";
+        Map<String, String> headers =
+                Map.of(
+                        "Authorization",
+                        "Bearer " + bearerToken,
+                        "client",
+                        CLIENT_NAME,
+                        "Content-Type",
+                        "application/fhir+ndjson");
+        try {
+            res = client.post(RS_WATERS_API_URL, headers, json);
+        } catch (HttpClientException e) {
+            throw new UnableToSendOrderException("Error POSTing the payload to ReportStream", e);
+        }
+
+        return res;
+    }
+
+    protected String requestToken() throws UnableToSendOrderException {
+        logger.logInfo("Requesting token from ReportStream");
+
+        String ourPrivateKey;
+        String token;
+
+        try {
+            ourPrivateKey = retrievePrivateKey();
+            String senderToken =
+                    jwt.generateToken(
+                            CLIENT_NAME,
+                            CLIENT_NAME,
+                            CLIENT_NAME,
+                            RS_DOMAIN_NAME,
+                            300,
+                            ourPrivateKey);
+            String body = composeRequestBody(senderToken);
+            String rsResponse = client.post(RS_AUTH_API_URL, RS_AUTH_API_HEADERS, body);
+            token = extractToken(rsResponse);
+        } catch (Exception e) {
+            throw new UnableToSendOrderException(
+                    "Error getting the API token from ReportStream", e);
+        }
+
+        // only cache our private key if we successfully authenticate to RS
+        cacheOurPrivateKeyIfNotCachedAlready(ourPrivateKey);
+
+        return token;
+    }
+
+    protected String retrievePrivateKey() throws SecretRetrievalException {
+        String key = cache.get(OUR_PRIVATE_KEY_ID);
+        if (key != null) {
+            return key;
+        }
+
+        key = secrets.getKey(OUR_PRIVATE_KEY_ID);
+
+        return key;
+    }
+
+    void cacheOurPrivateKeyIfNotCachedAlready(String privateKey) {
+        String key = cache.get(OUR_PRIVATE_KEY_ID);
+        if (key != null) {
+            return;
+        }
+
+        cache.put(OUR_PRIVATE_KEY_ID, privateKey);
+    }
+
+    protected String extractToken(String responseBody) throws FormatterProcessingException {
+        var value =
+                formatter.convertJsonToObject(
+                        responseBody, new TypeReference<Map<String, String>>() {});
+        return value.get("access_token");
+    }
+
+    protected String composeRequestBody(String senderToken) {
+        String scope = "flexion.*.report";
+        String grantType = "client_credentials";
+        String clientAssertionType = "urn:ietf:params:oauth:client-assertion-type:jwt-bearer";
+        return "scope="
+                + scope
+                + "&grant_type="
+                + grantType
+                + "&client_assertion_type="
+                + clientAssertionType
+                + "&client_assertion="
+                + senderToken;
+    }
+}
